ota_mechanism/requirements.txt
requests==2.27.1
psutil==5.9.0
azure-identity==1.13.0
azure-mgmt-sql==3.0.1
azure-storage-blob==12.16.0